fixed bug
if a user has already used the skript, which means an image and a container is available,
and the user executes run.sh again and forces a rebuild of the image and stops run.sh before
finishing the build process, there is no image available anymore, but a container.
If the user executes the skript again after that with just the option to pass a path, the skript will
build the image, but stop after that, because there's a container available.