Fix issue with relative path
If file is existing it is trying to execute the command using `ShellExecutionStrategy`, but if not we don't pass this command to any Strategy. My proposition is to check is this command has correct path format, and if so - execute it by `ShellExecutionStrategy`. If not - just go to the next available Strategy.